requirements.txt
# ================================================================
# Nexura Requirements
# ================================================================

# ── Django Core ───────────────────────────────────────────────────
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3

# ── Database ──────────────────────────────────────────────────────
psycopg2-binary==2.9.9   # PostgreSQL (optional — SQLite used by default)

# ── Cache & Queue ─────────────────────────────────────────────────
redis==5.0.3
celery==5.3.6
django-celery-beat==2.6.0
django-celery-results==2.5.1

# ── API Documentation ─────────────────────────────────────────────
drf-spectacular==0.27.1

# ── ML / AI ───────────────────────────────────────────────────────
xgboost==2.0.3
scikit-learn==1.4.2
joblib==1.3.2
pandas==2.2.1
numpy==1.26.4
prophet==1.1.5

# ── Payments ─────────────────────────────────────────────────────
razorpay==1.4.1

# ── Notifications ─────────────────────────────────────────────────
sendgrid==6.11.0
twilio==8.13.0
requests==2.31.0

# ── Documents & Storage ───────────────────────────────────────────
reportlab==4.1.0
qrcode==7.4.2
Pillow==10.2.0
boto3==1.34.69
django-storages==1.14.2
cryptography==42.0.5

# ── Production ────────────────────────────────────────────────────
gunicorn==21.2.0
whitenoise==6.6.0

# ── Dev / Testing ─────────────────────────────────────────────────
coverage==7.4.4
factory-boy==3.3.0


anthropic>=0.40.0          # Anthropic Claude — pip install anthropic
openai>=1.40.0             # OpenAI GPT-4o   — pip install openai
google-generativeai>=0.8.0 # Google Gemini   — pip install google-generativeai
